feat: add composition api，optimize order, combine taptap login & xd login, update icon
1. now order component has been split to wrap & instance which will be reusable in future
2. add composition api in order to reducing code complexity
3. update all icons in profile
4. remove tap login icon. tap user should use xd account for login
