+++ src/components/app/v2/order/OrderV2.scss
@@ -204,6 +204,7 @@
       line-height: 20px;
       &_active {
         color: #ff6600;
+        border: 1px solid #ff6600;
       }
     }
   }
