+++ src/components/app/v2/order/ProgressLine.tsx
@@ -1,8 +1,8 @@
 import { VNode } from 'vue';
-import { Component, Prop, Vue } from 'vue-property-decorator';
+import { Component, Vue } from 'vue-property-decorator';
 import * as tsx from 'vue-tsx-support';
 @Component
-export default class Limited extends tsx.Component<any> {
+export default class ProgressLine extends tsx.Component<any> {
   private render(): VNode {
     return <div></div>;
   }
